requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
reportlab==4.0.8
beautifulsoup4==4.12.2 